Housekeeping & Data changes
Have made all logs/data files now write under node /tmp/nias, easy to
then delete the whole lot for testing and easier to find things for
maintenance.

Have removed un-needed services ‘un-validator’ and moved privacy filter
view module to be in the web area with other sinatra services.

Have gone through all services and set class variables to be used for
service name - there was a conflict under high volume between two that
thought they were the same consumer so not all messages being processed.

Have downloaded a matched set of data from hits which is now in the
test_data directory, and which can be queried with sms once read in
