+++ sms/sms-redis.conf
@@ -184,7 +184,7 @@ dbfilename sms-dump.rdb
 # The Append Only File will also be created inside this directory.
 #
 # Note that you must specify a directory here, not a file name.
-dir '/tmp/redis'
+dir '/tmp/nias/redis'
 
 ################################# REPLICATION #################################
