--- test_data/schoolinfo.xml
@@ -1,122 +1,0 @@
-<messages>
-    <message>
-        <SchoolInfo RefId="dfe626b0-0e6e-4363-92a0-54ef1b190c6e">
-            <LocalId>01011234</LocalId>
-            <StateProvinceId>01011234</StateProvinceId>
-            <CommonwealthId>012345</CommonwealthId>
-            <SchoolName>Lincoln Secondary College</SchoolName>
-            <LEAInfoRefId>73648462888624AA5294BC6380173276</LEAInfoRefId>
-            <OtherLEA SIF_RefObject="LEAInfo">AA648462888624AA5294BC638017320B</OtherLEA>
-            <SchoolDistrict> Southern Metropolitan Region</SchoolDistrict>
-            <SchoolType>Pri/Sec</SchoolType>
-            <SchoolFocusList>
-                <SchoolFocus>01</SchoolFocus>
-                <SchoolFocus>03</SchoolFocus>
-            </SchoolFocusList>
-            <SchoolURL>http://www.lincolnsc.edu.vic.au</SchoolURL>
-            <PrincipalInfo>
-                <ContactName Type="LGL">
-                    <Title>Mr</Title>
-                    <FamilyName>Mason</FamilyName>
-                    <GivenName>Francis</GivenName>
-                    <MiddleName>Frank</MiddleName>
-                    <Suffix>Jr.</Suffix>
-                    <FullName>Mr Frank Mason Jr.</FullName>
-                </ContactName>
-                <ContactTitle>Senior School Principal</ContactTitle>
-            </PrincipalInfo>
-            <SchoolContactList>
-                <SchoolContact>
-                    <PublishInDirectory>Y</PublishInDirectory>
-                    <ContactInfo>
-                        <Name Type="LGL">
-                            <Title>Mr</Title>
-                            <FamilyName>Miller</FamilyName>
-                            <GivenName>James</GivenName>
-                            <MiddleName>Mark</MiddleName>
-                            <Suffix>Jr.</Suffix>
-                            <FullName>Mr James Mark Miller Jr.</FullName>
-                        </Name>
-                        <PositionTitle>Business Manager</PositionTitle>
-                        <Role>School Information Contact Point</Role>
-                        <Address Type="0123" Role="012B">
-                            <Street>
-                                <Line1>23 Nicholson Street</Line1>
-                            </Street>
-                            <City>Carnegie</City>
-                            <StateProvince>VIC</StateProvince>
-                            <Country>1101</Country>
-                            <PostalCode>3004</PostalCode>
-                            <GridLocation>
-                                <Latitude>23.9876</Latitude>
-                                <Longitude>-98.8765</Longitude>
-                            </GridLocation>
-                        </Address>
-                        <EmailList>
-                            <Email Type="01">jmiller@lsc.vic.edu.au</Email>
-                            <Email Type="02">jmiller@yahoo.com.au</Email>
-                        </EmailList>
-                        <PhoneNumberList>
-                            <PhoneNumber Type="0096">
-                                <Number>03 9637-2000</Number>
-                                <Extension>72345</Extension>
-                                <ListedStatus>Y</ListedStatus>
-                            </PhoneNumber>
-                        </PhoneNumberList>
-                    </ContactInfo>
-                </SchoolContact>
-            </SchoolContactList>
-            <PhoneNumberList>
-                <PhoneNumber Type="0096">
-                    <Number>03 9637-2000</Number>
-                </PhoneNumber>
-            </PhoneNumberList>
-            <SessionType>0827</SessionType>
-            <YearLevels>
-                <YearLevel>
-                    <Code>6</Code>
-                </YearLevel>
-                <YearLevel>
-                    <Code>7</Code>
-                </YearLevel>
-                <YearLevel>
-                    <Code>8</Code>
-                </YearLevel>
-                <YearLevel>
-                    <Code>9</Code>
-                </YearLevel>
-                <YearLevel>
-                    <Code>10</Code>
-                </YearLevel>
-                <YearLevel>
-                    <Code>11</Code>
-                </YearLevel>
-                <YearLevel>
-                    <Code>12</Code>
-                </YearLevel>
-            </YearLevels>
-            <ARIA>1.0</ARIA>
-            <OperationalStatus>O</OperationalStatus>
-            <FederalElectorate>216</FederalElectorate>
-            <Campus>
-                <SchoolCampusId>01</SchoolCampusId>
-                <CampusType>Camp</CampusType>
-                <AdminStatus>Y</AdminStatus>
-            </Campus>
-            <SchoolSector>NG</SchoolSector>
-            <IndependentSchool>Y</IndependentSchool>
-            <NonGovSystemicStatus>S</NonGovSystemicStatus>
-            <System>0003</System>
-            <ReligiousAffiliation>2171</ReligiousAffiliation>
-            <SchoolGeographicLocation>1</SchoolGeographicLocation>
-            <LocalGovernmentArea>Cardinia</LocalGovernmentArea>
-            <JurisdictionLowerHouse>Unknown</JurisdictionLowerHouse>
-            <SLA>205801452</SLA>
-            <SchoolCoEdStatus>C</SchoolCoEdStatus>
-            <SchoolGroupList>
-                <SchoolGroup>YVC</SchoolGroup>
-                <SchoolGroup>EastSec01</SchoolGroup>
-            </SchoolGroupList>
-        </SchoolInfo>
-    </message>
-</messages>
